Move form builder to own folder
